+++ collector/uptime_collector.go
@@ -1,6 +1,7 @@
 package collector
 
 import (
+	"regexp"
 	"strconv"
 	"strings"
 
@@ -18,7 +19,8 @@ var (
 
 func init() {
 	registerCollector("uptime", defaultEnabled, NewUptimeCollector)
-	uptimeDesc = prometheus.NewDesc(prefix+"uptime", "Displays how long the system has been running", labelnames, nil)
+	label_name_uptime := append(labelnames, "version")
+	uptimeDesc = prometheus.NewDesc(prefix+"uptime", "Displays how long the system has been running", label_name_uptime, nil)
 	loadLongtermDesc = prometheus.NewDesc(prefix+"load_longterm", "The average system load over a period of the last 15 minutes.", labelnames, nil)
 	loadMidtermDesc = prometheus.NewDesc(prefix+"load_midterm", "The average system load over a period of the last 5 minutes.", labelnames, nil)
 	loadShorttermDesc = prometheus.NewDesc(prefix+"load_shortterm", "The average system load over a period of the last 1 minutes.", labelnames, nil)
@@ -38,16 +40,22 @@ func (*uptimeCollector) Describe(ch chan<- *prometheus.Desc) {
 
 func (c *uptimeCollector) Collect(client *connector.SSHConnection, ch chan<- prometheus.Metric, labelvalue []string) error {
 	log.Debugln("uptime collector is starting")
+
 	results, err := client.RunCommand("uptime")
+	result_version, err := client.RunCommand("version")
 	if err != nil {
 		return err
 	}
+
 	var result []string = strings.Split(results, " ")
 	uptime, err := convertToSeconds(result[3], strings.Trim(result[5], ","))
 	loadLongterm, err := strconv.ParseFloat(strings.Trim(result[10], ","), 64)
 	loadMidterm, err := strconv.ParseFloat(strings.Trim(result[11], ","), 64)
 	loadShortterm, err := strconv.ParseFloat(strings.Trim(result[12], ",\n"), 64)
-	ch <- prometheus.MustNewConstMetric(uptimeDesc, prometheus.GaugeValue, uptime, labelvalue...)
+	re := regexp.MustCompile(`v\d+(\.\d+)*(\w)*`)
+	metric := re.FindString(result_version)
+	label_value_uptime := append(labelvalue, metric)
+	ch <- prometheus.MustNewConstMetric(uptimeDesc, prometheus.GaugeValue, uptime, label_value_uptime...)
 	ch <- prometheus.MustNewConstMetric(loadLongtermDesc, prometheus.GaugeValue, loadLongterm, labelvalue...)
 	ch <- prometheus.MustNewConstMetric(loadMidtermDesc, prometheus.GaugeValue, loadMidterm, labelvalue...)
 	ch <- prometheus.MustNewConstMetric(loadShorttermDesc, prometheus.GaugeValue, loadShortterm, labelvalue...)